requirements.txt
django_annoying==0.10.3
channels==0.17.3
Django==1.10.3
daphne==0.14.1
django_imagekit==3.3
dj-database-url==0.4.0
dj-static==0.0.6
Unipath==1.0
python-decouple==3
Pillow==3.4.2
psycopg2==2.6.1
whitenoise==3.2
dj-database-url==0.4.0
asgi_redis==1.0.0
asgiref==0.13
Twisted==15.5.0
txaio==2.6.0
zope.interface==4.1.3
haikunator==1.0.1
six==1.10.0
msgpack-python==0.4.7
uwsgi==2.0.14
djangorestframework==3.5.3
django-allauth==0.30.0



